requirements.txt
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
python-dotenv>=1.0
passlib[bcrypt]>=1.7
bcrypt<4.1
python-multipart>=0.0.9
reportlab>=4.0
pandas>=2.0
openpyxl>=3.1
